doc: added javadoc and fixed owner repository methods signature

+++ kotiki-app/service/src/main/java/ru/tolstov/services/OwnerServiceImpl.java
@@ -3,25 +3,25 @@ package ru.tolstov.services;
 import ru.tolstov.models.Owner;
 import ru.tolstov.repositories.OwnerRepository;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 public class OwnerServiceImpl implements OwnerService {
-    private OwnerRepository ownerRepository;
+    private final OwnerRepository ownerRepository;
 
     public OwnerServiceImpl(OwnerRepository ownerRepository) {
         this.ownerRepository = ownerRepository;
     }
 
     @Override
-    public Owner createOwner(String firstName, String lastName, Date birhdate) {
-        var owner = new Owner();
-        owner.setFirstName(firstName);
-        owner.setLastName(lastName);
-        owner.setBirthdate(birhdate);
-        ownerRepository.registerOwner(owner);
-
-        return owner;
+    public long createOwner(String firstName, String lastName, Date birhdate) {
+        return ownerRepository.registerOwner(owner -> {
+            owner.setFirstName(firstName);
+            owner.setLastName(lastName);
+            owner.setBirthdate(birhdate);
+            owner.setCats(new ArrayList<>());
+        });
     }
 
     @Override
@@ -30,7 +30,14 @@ public class OwnerServiceImpl implements OwnerService {
     }
 
     @Override
-    public void removeOwner(Owner owner) {
+    public void removeOwner(long ownerID) {
+        var owner = ownerRepository.getOwnerById(ownerID);
+        if (owner == null)
+            return;
+
+        if (!owner.getCats().isEmpty())
+            throw new RuntimeException("Cant remove owner while he has cats");
+
         ownerRepository.deleteOwner(owner);
     }
 }
